test-requirements.txt
-e .
defusedxml
flake8
flake8-import-order
requests-mock
coverage
pytest
pytest-cov
pytest-xdist
Cheetah3
psycopg2-binary
python-multilib
python-qpid-proton
